Move LTAN calculations to Earth
