+++ tests/tests_twobody/test_orbit.py
@@ -755,12 +755,6 @@ def test_heliosynchronous_orbit_enough_arguments():
     assert "Two parameters of (a, ecc, inc) are required" in excinfo.exconly()
 
 
-def test_heliosynchronous_orbit_without_earth():
-    with pytest.raises(NotImplementedError) as excinfo:
-        Orbit.heliosynchronous(Mars, a=800 * u.km + Mars.R, ecc=0 * u.one)
-    assert "Attractors other than Earth not supported yet" in excinfo.exconly()
-
-
 def test_heliosynchronous_orbit_inc():
     # Vallado, example 11-2a
     expected_ecc = 0 * u.one
